chore: move example files to better folders
