+++ examples/route-manager/openapi/manager/index.ts
@@ -1,7 +1,5 @@
 import { OpenAPIManager } from '@thequinndev/route-manager/openapi/manager'
-import { OpenApiSpecCompiler } from '@thequinndev/route-manager/openapi/compiler'
-import { apiDocumentationEndpoints, userEndpoints } from '../endpoints'
-import { writeFileSync } from 'fs';
+import { apiDocumentationEndpoints, userEndpoints } from '../../endpoints'
 
 const apiDocumentationDocument = OpenAPIManager({
     version: '3.0',
@@ -107,23 +105,5 @@ apiUserDocument.addEndpointGroup(userEndpoints)
     }
 })
 
-const compiler = OpenApiSpecCompiler({
-    version: '3.0',
-    specFile: {
-        openapi: '3.0.0',
-        info: {
-            
-            title: 'OpenAPIManager Example',
-            version: '1.0.0'
-        }
-    },
-    openApiManagers: [
-        apiDocumentationDocument,
-        apiUserDocument
-    ]
-})
-
-const apiSpec = compiler.build({
-    failOnError: true,
-})
-writeFileSync(__dirname + '/openapi.json', JSON.stringify(apiSpec, null, 2))
+export const ApiDocumentExample = apiDocumentationDocument
+export const UserDocumentExample = apiUserDocument
